Implicit conversion warnings.

+++ libs/openFrameworks/math/ofMatrix4x4.h
@@ -1042,7 +1042,7 @@ inline void ofMatrix4x4::rotate(float angle, float x, float y, float z){
 
 /// \brief Rotates this Matrix by the provided angle (in Radians) around an axis defined by the three values
 inline void ofMatrix4x4::rotateRad(float angle, float x, float y, float z){
-	postMultRotate(angle*RAD_TO_DEG,x,y,z);
+	postMultRotate(angle*static_cast<float>(RAD_TO_DEG),x,y,z);
 }
 
 /// \brief Translates this matrix by the provided amount
@@ -1071,7 +1071,7 @@ inline void ofMatrix4x4::glRotate(float angle, float x, float y, float z){
 }
 
 inline void ofMatrix4x4::glRotateRad(float angle, float x, float y, float z){
-	preMultRotate(ofQuaternion(angle*RAD_TO_DEG,ofVec3f(x,y,z)));
+	preMultRotate(ofQuaternion(angle*static_cast<float>(RAD_TO_DEG),ofVec3f(x,y,z)));
 }
 
 inline void ofMatrix4x4::glRotate(const ofQuaternion& q){
